pyproject.toml
[tool.poetry]
name = "kids-worksheet-generator"
version = "0.1.0"
description = ""
authors = ["Cristi Boboc <cristi@cbsoft.ro>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
Django = "^4.1.2"
psycopg2-binary = "^2.9.5"
pdfkit = "^1.0.0"
PyPDF2 = "^2.11.1"
django-cron = "^0.6.0"
django-axes = "^5.39.0"
mazelib = "^0.9.13"
python-datauri = "^1.1.0"
headless-pdfkit = "^0.1.5"
svglib = "^1.4.1"
selenium = "^4.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
